added motivation rspec test
