+++ .env
@@ -2,4 +2,4 @@ TOKEN= '1364171657:AAGcnycM0YLgWn1s2hsCkm-yfWw1g7-EOp4'
 TENOR_API_TOKEN = '4B2KMEXR10OX'
 GOOGLE_API= 'AIzaSyBag3db3WVCTe533vKkpGYAZbyRfTmz-BI'
 MERIAN_DIC= 'https://www.dictionaryapi.com/api/v3/references/learners/json/apple?key=e9adba98-cd77-4211-99d1-f5c44989dd84'
-DICTIONARY= 'e9adba98-cd77-4211-99d1-f5c44989dd84'
+MOTIVATION= 'https://programming-quotes-api.herokuapp.com/quotes/lang/en'
