+++ lib/karen.rb
@@ -61,24 +61,4 @@ class Karen
   end
 end
 
-class Motivation
-  attr_reader :motivate
 
-  def initialize
-    @motivate = message_request
-  end
-
-  def select_randomly
-    @motivate.sample
-  end
-
-  private
-
-  def message_request
-    url = 'https://programming-quotes-api.herokuapp.com/quotes/lang/en'
-    uri = URI(url)
-    response = Net::HTTP.get(uri)
-    response = JSON.parse(response)
-    response
-  end
-end
